initial theming attempt
The default (new) nextcloud theme/appearance with the rounded corners is somewhat controversial and since it serves no real purpose and fails to utilise the entire screen, an overhaul is in order. As such this docker-cloud theme will attempt to improve upon the classic look and feel and fix as many visual bugs/oversights as possible.